+++ OrangeDesignSystemDemo/Pages/SliderPage.swift
@@ -29,7 +29,7 @@ struct SliderPage: View {
 
     var body: some View {
         ScrollView {
-            Image("Controls")
+            Image("Slider")
                 .resizable()
                 .aspectRatio(contentMode: .fill)
             VStack(alignment: .leading, spacing: 20) {